refactor: 매일메일 api 관련 클래스 모듈 변경
