+++ maeilmail-api/src/main/java/maeilmail/api/QuestionApi.java
@@ -1,6 +1,8 @@
-package maeilmail.question;
+package maeilmail.api;
 
 import lombok.RequiredArgsConstructor;
+import maeilmail.question.QuestionQueryService;
+import maeilmail.question.QuestionSummary;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PathVariable;
