+++ maeilmail-api/src/main/java/maeilmail/api/StatisticsApi.java
@@ -1,7 +1,10 @@
-package maeilmail.statistics;
+package maeilmail.api;
 
 import java.time.LocalDate;
 import lombok.RequiredArgsConstructor;
+import maeilmail.statistics.DailySendReport;
+import maeilmail.statistics.StatisticsService;
+import maeilmail.statistics.SubscribeReport;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.RestController;
