+++ maeilmail-api/src/main/java/maeilmail/api/SubscribeQuestionApi.java
@@ -1,4 +1,4 @@
-package maeilmail.subscribe.api;
+package maeilmail.api;
 
 import lombok.RequiredArgsConstructor;
 import maeilmail.subscribe.query.SubscribeQuestionQueryService;
