+++ maeilmail-api/src/test/java/maeilmail/api/SubscribeQuestionApiTest.java
@@ -1,4 +1,4 @@
-package maeilmail.subscribe.api;
+package maeilmail.api;
 
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.junit.jupiter.api.Assertions.assertAll;
